Refactor hooks file into individual hooks
